fsmonitor: rewrite inotify support
File system change monitoring on Linux was implemented via the pyinotify
library, but this library has a cumbersome and limiting API.  In
particular when adding a watch, certain errors can occur due to race
conditions with simultaneous changes on the file system and are safe to
ignore, while others indicate an actual problem.  More specifically, the
limit on the number of watch descriptors may have been reached, which is
a fatal error.  pyinotify does not provide any way to distinguish
between these error cases, so git-cola was forced to treat any error
when adding a watch as fatal.  In addition, even though the underlying
inotify API supports a purely event-driven I/O model, pyinotify does not
support this model and requires polling with timeouts.

In this commit, add a new module, cola.inotify, that uses ctypes to
provide a thin wrapper around the inotify functions in libc   Rewrite
the cola.fsmonitor._InotifyThread class was to use the cola.inotify
module instead of pyinotify.  Now, properly handle the add_watch error
cases, ignoring those that are not fatal errors.  Change the monitoring
approach as well.  Now everything within the work tree is monitored
(except for the git dir).  This allows the creation and deletion of
untracked files to be monitored   Also properly monitor and account for
the creation and deletion of new directories.  Previously if a directory
was created after git-cola started and new files committed in that
directory, subsequent changes to those files would not cause git-cola to
refresh.  Also properly handle inotify queue overflows.  Finally, switch
to a fully-event driven I/O model, which means that the monitoring
thread no longer has to periodically wake up to check if it should stop
running.  Instead, it waits on both the inotify FD as well as the read
end of a pipe.  When the thread needs to be stopped, a byte is written
to the pipe, waking it up.

Signed-off-by: Daniel Harding <dharding@living180.net>