fsmonitor: eliminate timeout on Windows
Rework the Windows implementation of the file system monitoring code to
no longer us a timeout to check for the exit condition.  Instead, create
a separate Windows event object, and change the WaitForSingleObject call
to a WaitForMultipleObjects call that waits on both the
overlapped.hEvent as well as the control event.  To stop the thread, set
the _running flag to False and signal the control event to cause the
WaitForMultipleObjects call to return.

Signed-off-by: Daniel Harding <dharding@living180.net>