core: limit Windows subprocess hack to Python 2
In the core.start_command() function, there is a hack for Windows to set
the cwd variable to None if it equals the current directory, due to the
fact that Python 2 does not handle non-ASCII characters being passed as
the cwd to subprocess.  This is not an issue on Python 3, so limit this
hack to Python 2.

Signed-off-by: Daniel Harding <dharding@living180.net>