git: remove unnecessary code
In the Git.execute() method, a check was being made for an empty _cwd,
and setting it to the current directory if that is the case.  This is
unnecessary as passing cwd=None to core.run_command() has the effect as
running in the current directory.  Remove the unnecessary check.

Signed-off-by: Daniel Harding <dharding@living180.net>