fsmonitor: ignore directory changes
Since git only tracks files and not directories, update the file system
change monitoring code to ignore changes to directories.  This will help
prevent unnecessary refreshes e.g. when directories are added or removed
as part of a build process.  It could potentially lead to some file
system changes not triggering a refresh even if tracked files are
affected (e.g. if a directory containing tracked files is moved or
renamed), but I think this is an acceptable trade-off.

Signed-off-by: Daniel Harding <dharding@living180.net>